# docs/swagger/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_api.abortActivityRequest:
    properties:
      reason:
        type: string
    type: object
  internal_api.activityDetailResponse:
    properties:
      activity:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.Activity'
      events:
        items:
          $ref: '#/definitions/jpcorrect-backend_internal_domain.Event'
        type: array
    type: object
  internal_api.createActivityRequest:
    properties:
      announce_topic_id:
        type: string
      mode:
        type: string
      practice_at:
        type: string
      reporter_user_ids:
        items:
          type: string
        type: array
      review_at:
        type: string
      theme:
        type: string
    required:
    - practice_at
    type: object
  internal_api.updateActivityRequest:
    properties:
      mode:
        type: string
      practice_at:
        type: string
      reporter_user_ids:
        items:
          type: string
        type: array
      review_at:
        type: string
      theme:
        type: string
    type: object
  jpcorrect-backend_internal_domain.AIStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - AIStatusPending
    - AIStatusProcessing
    - AIStatusCompleted
    - AIStatusFailed
  jpcorrect-backend_internal_domain.Activity:
    properties:
      aborted_at:
        type: string
      aborted_by_user_id:
        type: string
      aborted_reason:
        type: string
      activity_id:
        type: string
      ai_status:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.AIStatus'
      announce_topic_id:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      guild_id:
        type: string
      mode:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.ActivityMode'
      sequence_number:
        type: integer
      status:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.ActivityStatus'
      theme:
        type: string
      updated_at:
        type: string
      youtube_status:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.YoutubeStatus'
      youtube_url:
        type: string
    type: object
  jpcorrect-backend_internal_domain.ActivityMode:
    enum:
    - report
    - conversation
    type: string
    x-enum-varnames:
    - ActivityModeReport
    - ActivityModeConversation
  jpcorrect-backend_internal_domain.ActivityStatus:
    enum:
    - pending_practice
    - in_practice
    - analyzing
    - in_feedback
    - in_review
    - done
    - aborted
    type: string
    x-enum-varnames:
    - ActivityStatusPendingPractice
    - ActivityStatusInPractice
    - ActivityStatusAnalyzing
    - ActivityStatusInFeedback
    - ActivityStatusInReview
    - ActivityStatusDone
    - ActivityStatusAborted
  jpcorrect-backend_internal_domain.Event:
    properties:
      activity_id:
        type: string
      actual_duration:
        type: number
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      event_id:
        type: string
      expected_duration:
        type: number
      mode:
        $ref: '#/definitions/jpcorrect-backend_internal_domain.EventMode'
      note:
        type: string
      record_link:
        type: string
      recording_ended_at:
        type: string
      recording_started_at:
        type: string
      recording_started_by:
        type: string
      start_time:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  jpcorrect-backend_internal_domain.EventMode:
    enum:
    - report
    - conversation
    - review
    type: string
    x-enum-varnames:
    - EventModeReport
    - EventModeConversation
    - EventModeReview
...
  jpcorrect-backend_internal_domain.YoutubeStatus:
    enum:
    - pending
    - uploading
    - completed
    - failed
    type: string
    x-enum-varnames:
    - YoutubeStatusPending
    - YoutubeStatusUploading
    - YoutubeStatusCompleted
    - YoutubeStatusFailed
host: localhost:8080
info:
  contact: {}
  description: Japanese language correction platform backend API
  title: jpcorrect API
  version: "1.0"
paths:
  /v1/activities/{id}:
    get:
      consumes:
      - application/json
      description: Get an activity by ID with its events
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.activityDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an activity
      tags:
      - activities
    put:
      consumes:
      - application/json
      description: Update an activity's time, mode, theme, and corresponding events
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: Activity update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.updateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.activityDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an activity
      tags:
      - activities
  /v1/activities/{id}/abort:
    post:
      consumes:
      - application/json
      description: Abort an activity with an optional reason
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: Abort data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.abortActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jpcorrect-backend_internal_domain.Activity'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Abort an activity
      tags:
      - activities
...
  /v1/guilds/{id}/activities:
    get:
      consumes:
      - application/json
      description: List activities for a guild with pagination and optional status
        filter
      parameters:
      - description: Guild ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jpcorrect-backend_internal_domain.Activity'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List guild activities
      tags:
      - activities
    post:
      consumes:
      - application/json
      description: Create a new activity with practice and review events for a guild
      parameters:
      - description: Guild ID
        in: path
        name: id
        required: true
        type: string
      - description: Activity data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.createActivityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.activityDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an activity
      tags:
      - activities
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"